requirements.txt
dvc==2.6.4
joblib==1.0.1
jupyter==1.0.0
jupyter_contrib_nbextensions==0.5.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
pytest==6.2.4
python-box==5.4.1
pyyaml==5.4.1
scikit-learn==0.24.2
scipy==1.7.1
tqdm==4.62.2